Phase 2 cleanup: remove dead code, split EEP layers (#397)
* Remove PriceTableWorkspace (Arrow IPC persistence)

Not used by core pricing path — only exposed via Python bindings.
Removes ~1280 lines including Arrow dependency, serialization
logic, and associated tests.


* Remove PriceSurface concept

Only PriceTable<Inner> ever satisfied it. InterpolatedIVSolver
now uses a plain template parameter constrained by usage, not
a concept that duplicated PriceTable's interface.


* Split EEPSurfaceAdapter into TransformLeaf + EEPLayer

TransformLeaf<Interp, Xform> handles coordinate transform,
raw interpolation, and K/K_ref scaling. EEPLayer<Leaf, EEP>
adds European price/vega on top.

Segmented leaves use TransformLeaf directly (no EEP wrapper).
Standard leaves compose EEPLayer<TransformLeaf<...>, AnalyticalEEP>.
This eliminates IdentityEEP and prepares the seam for 4D→3D
dimension reduction (issue #382).


* Split AdaptiveGridBuilder into modular components (#398)

* Add design for splitting adaptive_grid_builder


* Add implementation plan for splitting AGB

9-task plan to replace the 2296-line AdaptiveGridBuilder
god class with free functions per backend, keeping shared
refinement infrastructure in adaptive_refinement.{hpp,cpp}.


* Extract shared refinement infra from AGB

Move shared types (RefinementContext, RefinementResult,
ErrorBins, InitialGrids, etc.) and 14 helper functions
(run_refinement, seed_grid, linspace, etc.) into
adaptive_refinement.{hpp,cpp}.

Old adaptive_grid_builder.cpp now uses these via the
shared header. Both targets coexist and compile.


* Extract B-spline adaptive builder into free fns

Move B-spline-specific code from AdaptiveGridBuilder into
bspline_adaptive.{hpp,cpp} with typed results:
- build_adaptive_bspline() replaces builder.build()
- build_adaptive_bspline_segmented() replaces builder.build_segmented()

Results use typed surface fields instead of std::any.
SliceCache is now local to each build call.


* Extract Chebyshev adaptive builder into free fns

Move Chebyshev-specific code from AdaptiveGridBuilder into
chebyshev_adaptive.{hpp,cpp} with typed results:
- build_adaptive_chebyshev() replaces builder.build_chebyshev()
- build_adaptive_chebyshev_segmented() replaces
  builder.build_segmented_chebyshev()

Old adaptive_grid_builder.cpp reduced to empty class shell.


* Update consumers to use free function builders

Replace AdaptiveGridBuilder class usage with typed free
function calls in interpolated_iv_solver, tests, and
benchmarks. Removes std::any_cast — results are now typed.


* Delete AdaptiveGridBuilder and clean up types

- Move SegmentedAdaptiveConfig to adaptive_grid_types.hpp
- Remove AdaptiveResult and SegmentedAdaptiveResult (replaced
  by typed results per backend)
- Delete adaptive_grid_builder.{hpp,cpp} and BUILD target
- Update all includes and deps across consumers


* Use FD American vega for all error metrics

All adaptive refinement paths now use make_fd_vega_error_fn which
computes vega via central finite differences on the American PDE
solver. This replaces the previous BS European vega approximation
which was inaccurate for American options.

Also removes dead compute_error_metric, unused BS analytics
includes/deps, and the empty bspline_interpolant.hpp shim.


* Fix review issues in IV solver factory

- Guard k_ref_count==1 in build_multi_kref_manual (divide-by-zero)
- Guard empty IVGrid vectors in extract_bounds (UB on empty range)
- Wire Chebyshev+segmented through AnyIVSolver via FDVegaLeaf
  adapter instead of silently falling back to B-spline


* Extract shared PDE cache infrastructure

Factor out common (sigma, rate) quantized-key logic into
PairCacheCore<PairValue> template in pde_cache.hpp. B-spline
and Chebyshev caches become thin wrappers with consistent
naming: BSplinePDECache (1e6 scale) and ChebyshevPDECache
(1e12 scale).


* Fix review issues in Chebyshev factory paths

- Guard chebyshev_nodes() against num_pts < 2 (divide-by-zero)
- Add num_pts validation in build_chebyshev_table()
- Derive tau_min from maturity instead of hard-coding 0.01
- Add Chebyshev integration tests (continuous + segmented)


* Guard k_ref_count < 1 in manual segmented path

Prevents pathological allocation when negative K_ref_count
casts to huge size_t via reserve().


---------


---------
