+++ benchmarks/BUILD.bazel
@@ -180,11 +180,13 @@ cc_binary(
         "//src/option:iv_solver",
         "//src/option:interpolated_iv_solver",
         "//src/option:option_grid",
-        "//src/option/table:adaptive_grid_builder",
+        "//src/option/table/bspline:bspline_adaptive",
         "//src/option/table/bspline:bspline_builder",
         "//src/option/table/bspline:bspline_surface",
         "//src/option/table/bspline:bspline_segmented_builder",
         "//src/option/table/bspline:bspline_tensor_accessor",
+        "//src/option/table/chebyshev:chebyshev_adaptive",
+        "//src/option/table/chebyshev:chebyshev_surface",
         "//src/math:bspline_nd_separable",
         "@google_benchmark//:benchmark",
     ],
@@ -465,7 +467,8 @@ cc_binary(
         "//src/option:american_option_batch",
         "//src/option:iv_solver",
         "//src/option:interpolated_iv_solver",
-        "//src/option/table:adaptive_grid_builder",
+        "//src/option/table:adaptive_grid_types",
+        "//src/option/table/chebyshev:chebyshev_adaptive",
         "//src/option/table/chebyshev:chebyshev_table_builder",
     ],
     tags = ["manual"],
@@ -488,7 +491,7 @@ cc_binary(
         "//src/option:american_option_batch",
         "//src/option:european_option",
         "//src/option:interpolated_iv_solver",
-        "//src/option/table:adaptive_grid_builder",
+        "//src/option/table/bspline:bspline_adaptive",
         "//src/option/table/bspline:bspline_surface",
         "//src/option/table/bspline:bspline_builder",
         "//src/option/table/bspline:bspline_tensor_accessor",
