+++ src/option/BUILD.bazel
@@ -153,11 +153,12 @@ cc_library(
         ":option_spec",
         "//src/option/table/bspline:bspline_tensor_accessor",
         "//src/option/table/bspline:bspline_surface",
-        "//src/option/table:price_surface_concept",
+        "//src/option/table:price_table",
         "//src/option/table:adaptive_grid_types",
-        "//src/option/table:adaptive_grid_builder",
+        "//src/option/table/bspline:bspline_adaptive",
         "//src/option/table/bspline:bspline_builder",
         "//src/option/table/bspline:bspline_segmented_builder",
+        "//src/option/table/chebyshev:chebyshev_adaptive",
         "//src/option/table/chebyshev:chebyshev_surface",
         "//src/option/table/chebyshev:chebyshev_table_builder",
         "//src/math:root_finding",
