+++ src/option/table/bspline/BUILD.bazel
@@ -1,9 +1,10 @@
 # SPDX-License-Identifier: MIT
 cc_library(
-    name = "bspline_slice_cache",
-    hdrs = ["bspline_slice_cache.hpp"],
+    name = "bspline_pde_cache",
+    hdrs = ["bspline_pde_cache.hpp"],
     deps = [
         "//src/option:american_option_result",
+        "//src/option/table:pde_cache",
     ],
     visibility = ["//visibility:public"],
     strip_include_prefix = "/src/option/table/bspline",
@@ -13,16 +14,13 @@ cc_library(
 cc_library(
     name = "bspline_surface",
     srcs = ["bspline_surface.cpp"],
-    hdrs = [
-        "bspline_surface.hpp",
-        "bspline_interpolant.hpp",
-    ],
+    hdrs = ["bspline_surface.hpp"],
     deps = [
         "//src/option:option_spec",
         "//src/option/table:price_table",
-        "//src/option/table:eep_surface_adapter",
+        "//src/option/table:transform_leaf",
+        "//src/option/table:eep_layer",
         "//src/option/table:analytical_eep",
-        "//src/option/table:identity_eep",
         "//src/option/table:split_surface",
         "//src/option/table:tau_segment_split",
         "//src/option/table:multi_kref_split",
@@ -72,22 +70,6 @@ cc_library(
 )
 
 cc_library(
-    name = "bspline_workspace",
-    srcs = ["bspline_workspace.cpp"],
-    hdrs = ["bspline_workspace.hpp"],
-    deps = [
-        "//src/support:error_types",
-        "//src/math:bspline_basis",
-        "//src/support:crc64",
-        "//third_party/arrow",
-        "@mdspan//:mdspan",
-    ],
-    visibility = ["//visibility:public"],
-    strip_include_prefix = "/src/option/table/bspline",
-    include_prefix = "mango/option/table/bspline",
-)
-
-cc_library(
     name = "bspline_tensor_accessor",
     hdrs = ["bspline_tensor_accessor.hpp"],
     deps = [
@@ -111,6 +93,42 @@ cc_library(
         "//src/option:american_option",
         "//src/option:dividend_utils",
         "//src/option:grid_spec_types",
+        "//src/support:error_types",
+    ],
+    visibility = ["//visibility:public"],
+    strip_include_prefix = "/src/option/table/bspline",
+    include_prefix = "mango/option/table/bspline",
+)
+
+cc_library(
+    name = "bspline_adaptive",
+    srcs = ["bspline_adaptive.cpp"],
+    hdrs = ["bspline_adaptive.hpp"],
+    copts = [
+        "-fopenmp",
+        "-pthread",
+    ],
+    linkopts = [
+        "-fopenmp",
+        "-pthread",
+    ],
+    deps = [
+        ":bspline_builder",
+        ":bspline_segmented_builder",
+        ":bspline_pde_cache",
+        ":bspline_surface",
+        ":bspline_tensor_accessor",
+        "//src/math:cubic_spline_solver",
+        "//src/math:latin_hypercube",
+        "//src/option:american_option_batch",
+        "//src/option:dividend_utils",
+        "//src/option:option_grid",
+        "//src/option/table:adaptive_grid_types",
+        "//src/option/table:adaptive_refinement",
+        "//src/option/table:eep_decomposer",
+        "//src/option/table:multi_kref_split",
+        "//src/option/table:split_surface",
+        "//src/pde/core:time_domain",
         "//src/support:error_types",
     ],
     visibility = ["//visibility:public"],
