--- src/option/table/bspline/bspline_interpolant.hpp
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: MIT
-#pragma once
-
-// SharedBSplineInterp is now defined in bspline_surface.hpp.
-// This header is kept for backward compatibility.
-#include "mango/option/table/bspline/bspline_surface.hpp"
