+++ src/option/table/chebyshev/BUILD.bazel
@@ -8,8 +8,8 @@ cc_library(
         "//src/math/chebyshev:tucker_tensor",
         "//src/option/table:price_table",
         "//src/option/table:analytical_eep",
-        "//src/option/table:identity_eep",
-        "//src/option/table:eep_surface_adapter",
+        "//src/option/table:eep_layer",
+        "//src/option/table:transform_leaf",
         "//src/option/table:standard_transform_4d",
     ],
     visibility = ["//visibility:public"],
@@ -18,9 +18,44 @@ cc_library(
 )
 
 cc_library(
-    name = "pde_slice_cache",
-    hdrs = ["pde_slice_cache.hpp"],
-    deps = ["//src/math:cubic_spline_solver"],
+    name = "chebyshev_pde_cache",
+    hdrs = ["chebyshev_pde_cache.hpp"],
+    deps = [
+        "//src/math:cubic_spline_solver",
+        "//src/option/table:pde_cache",
+    ],
+    visibility = ["//visibility:public"],
+    strip_include_prefix = "/src/option/table/chebyshev",
+    include_prefix = "mango/option/table/chebyshev",
+)
+
+cc_library(
+    name = "chebyshev_adaptive",
+    srcs = ["chebyshev_adaptive.cpp"],
+    hdrs = ["chebyshev_adaptive.hpp"],
+    copts = [
+        "-fopenmp",
+        "-pthread",
+    ],
+    linkopts = [
+        "-fopenmp",
+        "-pthread",
+    ],
+    deps = [
+        ":chebyshev_surface",
+        ":chebyshev_pde_cache",
+        "//src/math/chebyshev:chebyshev_nodes",
+        "//src/option:american_option_batch",
+        "//src/option:dividend_utils",
+        "//src/option:option_grid",
+        "//src/option/table:adaptive_grid_types",
+        "//src/option/table:adaptive_refinement",
+        "//src/option/table:eep_decomposer",
+        "//src/option/table:multi_kref_split",
+        "//src/option/table:split_surface",
+        "//src/option/table:tau_segment_split",
+        "//src/support:error_types",
+    ],
     visibility = ["//visibility:public"],
     strip_include_prefix = "/src/option/table/chebyshev",
     include_prefix = "mango/option/table/chebyshev",
