+++ tests/chebyshev_surface_test.cc
@@ -4,7 +4,6 @@
 #include "mango/option/table/chebyshev/chebyshev_surface.hpp"
 #include "mango/option/table/chebyshev/chebyshev_table_builder.hpp"
 #include "mango/option/american_option.hpp"
-#include "mango/option/table/price_surface_concept.hpp"
 #include "mango/option/table/surface_concepts.hpp"
 
 using namespace mango;
@@ -12,8 +11,6 @@ using namespace mango;
 // Static assertions
 static_assert(SurfaceInterpolant<ChebyshevInterpolant<4, TuckerTensor<4>>, 4>);
 static_assert(SurfaceInterpolant<ChebyshevInterpolant<4, RawTensor<4>>, 4>);
-static_assert(PriceSurface<ChebyshevSurface>);
-static_assert(PriceSurface<ChebyshevRawSurface>);
 
 TEST(ChebyshevSurfaceTest, ConstructAndQuery) {
     Domain<4> domain{
@@ -26,11 +23,10 @@ TEST(ChebyshevSurfaceTest, ConstructAndQuery) {
         [](std::array<double, 4>) { return 0.05; },
         domain, num_pts, 1e-8);
 
-    ChebyshevLeaf leaf(
-        std::move(interp),
-        StandardTransform4D{},
-        AnalyticalEEP(OptionType::PUT, 0.02),
-        100.0);
+    ChebyshevTransformLeaf tleaf(
+        std::move(interp), StandardTransform4D{}, 100.0);
+    ChebyshevLeaf leaf(std::move(tleaf),
+        AnalyticalEEP(OptionType::PUT, 0.02));
 
     SurfaceBounds bounds{
         .m_min = -0.5, .m_max = 0.5,
