+++ tests/error_attribution_test.cc
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 #include <gtest/gtest.h>
-#include "mango/option/table/bspline/bspline_slice_cache.hpp"
+#include "mango/option/table/adaptive_refinement.hpp"
 
 namespace mango {
 namespace {
