--- tests/price_surface_concept_test.cc
@@ -1,15 +1,0 @@
-// SPDX-License-Identifier: MIT
-#include <gtest/gtest.h>
-#include "mango/option/table/price_surface_concept.hpp"
-#include "mango/option/table/bspline/bspline_surface.hpp"
-
-using namespace mango;
-
-static_assert(PriceSurface<BSplinePriceTable>,
-    "BSplinePriceTable must satisfy PriceSurface concept");
-
-TEST(PriceSurfaceConceptTest, BSplinePriceTableSatisfiesConcept) {
-    // Compile-time check above is the real test.
-    // This test exists so the test binary runs.
-    SUCCEED();
-}
